factored out auth and made meterreadings SUPER FAST
